# docs/openapi.json
{
  "components": {
    "schemas": {
      "ChatCompletionRequest": {
        "additionalProperties": true,
        "description": "`POST /v1/chat/completions` 요청 본문.\n\nAttributes:\n    model: Bedrock 모델 ID 또는 추론 프로파일 ID.\n    messages: 대화 이력.\n    max_tokens: 최대 출력 토큰. OpenAI 의 구 필드명.\n    max_completion_tokens: 최대 출력 토큰. OpenAI 의 신 필드명.\n        둘 다 오면 이쪽을 우선한다.\n    temperature: 샘플링 온도.\n    top_p: 누적 확률 절단값.\n    stop: 정지 문자열 또는 목록.\n    stream: SSE 스트리밍 여부.\n    n: 생성할 후보 수. 1만 지원한다.\n    user: 호출자 식별 문자열. 사용량은 API 키로 귀속되므로 무시한다.",
        "properties": {
          "max_completion_tokens": {
            "anyOf": [
              {
                "maximum": 200000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Completion Tokens"
          },
          "max_tokens": {
            "anyOf": [
              {
                "maximum": 200000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          },
          "model": {
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "n": {
            "default": 1,
            "maximum": 1.0,
            "minimum": 1.0,
            "title": "N",
            "type": "integer"
          },
          "stop": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop"
          },
          "stream": {
            "default": false,
            "title": "Stream",
            "type": "boolean"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "top_p": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top P"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          }
        },
        "required": [
          "model",
          "messages"
        ],
        "title": "ChatCompletionRequest",
        "type": "object"
      },
      "ChatMessage": {
        "additionalProperties": true,
        "description": "대화 메시지 한 건.\n\nAttributes:\n    role: `system`, `developer`, `user`, `assistant` 중 하나.\n    content: 문자열 또는 텍스트 조각 배열.\n    name: OpenAI 호환용 선택 필드. 변환에는 쓰지 않는다.",
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/ContentPart"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
      "ContentPart": {
        "additionalProperties": true,
        "description": "멀티모달 메시지의 텍스트 조각.\n\nAttributes:\n    type: 조각 종류. 현재는 `text` 만 처리한다.\n    text: 텍스트 내용.",
        "properties": {
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "type": {
            "default": "text",
            "title": "Type",
            "type": "string"
          }
        },
        "title": "ContentPart",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Amazon Bedrock 앞단의 OpenAI 호환 게이트웨이.\n\n- `POST /v1/chat/completions` — OpenAI Chat Completions 호환 (스트리밍 지원)\n- `GET /v1/models` — 호출 가능한 모델 목록\n- `/admin/*` — 계정·팀·사용자·API 키 관리 (`X-Admin-Token` 필요)\n- `/analytics/*` — 계정·팀·사용자·모델별 사용량 집계 (`X-Admin-Token` 필요)\n- `/ui` — 모니터링 대시보드",
    "title": "LLM Gateway",
    "version": "1.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "description": "채팅 완성을 수행한다.\n\nArgs:\n    payload: OpenAI 형식 요청 본문.\n    services: 서비스 컨테이너.\n    authorization: `Bearer <api-key>` 헤더.\n    x_request_id: 클라이언트가 지정한 요청 ID. 재시도 시 같은 값을\n        보내면 사용량이 중복 집계되지 않는다.\n\nReturns:\n    비스트리밍이면 OpenAI 형식 응답 딕셔너리, 스트리밍이면\n    `StreamingResponse`.\n\nRaises:\n    GatewayError: 인증·권한·예산·업스트림 오류가 발생한 경우.",
        "operationId": "chat_completions_v1_chat_completions_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Chat Completions V1 Chat Completions Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Completions",
        "tags": [
          "openai-compat"
        ]
      }
    },
    "/v1/models": {
      "get": {
        "description": "호출 가능한 모델 목록을 OpenAI 형식으로 반환한다.\n\n키에 허용 모델 목록이 설정돼 있으면 그 목록으로 제한한다. 클라이언트가\n쓸 수 없는 모델을 보여주면 선택 후 403 을 받게 되어 혼란스럽다.\n\nArgs:\n    services: 서비스 컨테이너.\n    authorization: `Bearer <api-key>` 헤더.\n\nReturns:\n    OpenAI 형식의 모델 목록.\n\nRaises:\n    AuthenticationError: API 키가 유효하지 않은 경우.",
        "operationId": "list_models_v1_models_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Models V1 Models Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Models",
        "tags": [
          "openai-compat"
        ]
      }
    }
  }
}